Fixed Numerous Issues
Changed Background, split into multiple files, added auto scaling
